+++ PF/arquivos_disponibilizados/Player.cc
@@ -10,7 +10,7 @@
 
 __BEGIN_API
 
-Player::Player(int size, double speed, int maxLife, Point position, ALLEGRO_COLOR missileColor, ALLEGRO_COLOR laserColor, int missileDelay, int laserDelay) : Spaceship(size, speed, maxLife, position, missileColor, laserColor, missileDelay, laserDelay) {
+Player::Player(int size, double speed, int maxLife, Point position, ALLEGRO_COLOR missileColor, ALLEGRO_COLOR laserColor, double missileDelay, double laserDelay) : Spaceship(size, speed, maxLife, position, missileColor, laserColor, missileDelay, laserDelay) {
   _missileMovement = Vector(500, 0);
   _laserMovement = Vector(500, 0);
 }
